requirements.txt
autopep8==1.6.0
boto3==1.23.7
botocore==1.26.7
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
